+++ composer.json
@@ -20,7 +20,7 @@
         "ext-json": "*",
         "psr/log": "^1",
         "psr/http-message": "^1",
-        "psr/http-client": "@dev",
+        "psr/http-client": "^0.3 || 1",
         "psr/http-factory": "^1.0",
         "fig/http-message-util": "^1.1",
         "http-interop/http-factory-guzzle": "^1.0"
