+++ src/Feature/Contacts/ContactsHttpFeature.php
@@ -8,11 +8,7 @@ use Fig\Http\Message\RequestMethodInterface;
 use Smsapi\Client\Feature\Contacts\Fields\ContactsFieldsFeature;
 use Smsapi\Client\Feature\Contacts\Fields\ContactsFieldsHttpFeature;
 use Smsapi\Client\Feature\Data\DataFactoryProvider;
-use Smsapi\Client\Infrastructure\Request\RequestBuilder;
-use Smsapi\Client\Infrastructure\Request\RequestBuilderFactory;
-use Smsapi\Client\Infrastructure\Request\RestRequestBuilder;
 use Smsapi\Client\Infrastructure\Request\RestRequestBuilderFactory;
-use Smsapi\Client\Infrastructure\RequestExecutor\RequestExecutor;
 use Smsapi\Client\Infrastructure\RequestExecutor\RestRequestExecutor;
 use Smsapi\Client\Feature\Contacts\Bag\CreateContactBag;
 use Smsapi\Client\Feature\Contacts\Bag\DeleteContactBag;
